move logger definitions to cpp
